fix(packages): keep sanitizer dependency public
