+++ packages/helpers/package.json
@@ -6,11 +6,13 @@
     "@genfeedai/interfaces": "workspace:*",
     "@genfeedai/pricing": "workspace:*",
     "clsx": "2.1.1",
+    "sanitize-html": "2.17.6",
     "tailwind-merge": "3.6.0",
     "ts-jsonapi": "2.1.3"
   },
   "description": "Shared helpers for Genfeed clients and services",
   "devDependencies": {
+    "@types/sanitize-html": "2.16.1",
     "@types/file-saver": "2.0.7",
     "@types/react": "19.2.17",
     "@types/react-dom": "19.2.3",
