+++ packages/ui/package.json
@@ -9,7 +9,6 @@
     "@genfeedai/helpers": "workspace:*",
     "@genfeedai/interfaces": "workspace:*",
     "@genfeedai/pricing": "workspace:*",
-    "@genfeedai/utils": "workspace:*",
     "@hookform/resolvers": "5.4.0",
     "@shipshitdev/ui": "0.8.3",
     "@tanstack/react-query": "5.101.2",
@@ -45,6 +44,7 @@
     "@genfeedai/models": "workspace:*",
     "@genfeedai/props": "workspace:*",
     "@genfeedai/services": "workspace:*",
+    "@genfeedai/utils": "workspace:*",
     "@tanstack/react-query-devtools": "5.101.2",
     "@testing-library/jest-dom": "6.9.1",
     "@testing-library/react": "16.3.2",
